+++ packages/flutter_tvos/lib/src/tvos_info_platform.dart
@@ -1,0 +1,6 @@
+// Copyright 2026 The FlutterTV Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+export 'tvos_info_platform_web.dart'
+    if (dart.library.io) 'tvos_info_platform_native.dart';
