+++ packages/flutter_tvos/lib/src/tvos_info_platform_native.dart
@@ -1,0 +1,7 @@
+// Copyright 2026 The FlutterTV Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+import 'dart:io' show Platform;
+
+bool get isApple => Platform.isIOS || Platform.isMacOS;
