+++ packages/flutter_tvos/lib/src/tvos_info_platform_web.dart
@@ -1,0 +1,6 @@
+// Copyright 2026 The FlutterTV Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+// Web stub — dart:io is not available on Web.
+bool get isApple => false;
